+++ packages/lib/src/propTransform/index.ts
@@ -1,4 +1,6 @@
 export { PropTransform, propTransform, PropTransformDecorator } from "./propTransform"
+export * from "./transformArrayAsMap"
 export * from "./transformArrayAsSet"
+export * from "./transformObjectAsMap"
 export * from "./transformStringAsDate"
 export * from "./transformTimestampAsDate"