+++ packages/lib/src/wrappers/arrayAsSetWrapper.ts
@@ -83,6 +83,8 @@ export class ArrayAsSet<V> implements Set<V> {
 }
 
 /**
+ * @deprecated Consider using `prop_setArray` or `tProp_setArray` instead.
+ *
  * Returns a wrapper that wraps an observable array `V[]`
  * into a set alike interface.
  *