+++ packages/site/doczrc.js
@@ -15,7 +15,7 @@ export default {
     "Root Stores",
     "Snapshots",
     "Patches",
-    "Maps & Sets",
+    "Maps, Sets, Dates",
     {
       name: "Action Middlewares",
       menu: [